Create interfaces folder, refactor to fit.
