+++ papermage_components/interfaces/token_classification_predictor.py
